extracted wire data into its own class. making a ubwiretools which provides us with some wire geometry functions
