+++ app/PMTWeights/PMTWireWeights.h
@@ -41,9 +41,6 @@ namespace larcv {
       int femch;
       float pmtpos[32][3];
       
-      TTree* fWireTree;
-      std::map<int,larcv::pmtweights::WireData> m_WireData; // key is plane ID, value is class with wire info
-
       // The object we end up filling
       std::map<int, larcv::Image2D> planeWeights; // index is the plane number
